seperate Component to Add user to a team
